Add Target MySQL dsn to
Trying to add the mysql url target to the `mysql_up` metric. With apologies for my first attempt go code.

In a scenario where the mysql scraper is separated from the database server, there is no straightforward way to determine which mysql instance is associated with the exporter. AFAIK, the other option here is to manually add labels to prometheus scraper targets, but that gets pretty messy as envs proliferate.

Does this raise any security issues? From [here](https://github.com/go-sql-driver/mysql) it looks like it has the potential to add the whole DSN @ `username:password@protocol(address)/dbname?param=value` which is no good. Is there a way to use slice to omit user:pass?